[CDF-14918] Base extensions class (#132)
Essentially just pulls the core code out of the MQTT and REST extensions and creates a common base extractor containing queues and the `handle_output` method. It isn't ideal to redefine types in all extension libraries, and it might be nice to have a place to put common behavior in the future.