Retry on requests.exceptions.ConnectionError (again) (#282)
In a previous PR, we switched from retrying on requests connection
errors to the internal Cognite exceptions that wrapped these. This
fixed some of the issues encountered by the customer (SLB).

But, in some corner cases, such as for authorization, you can still
get an unwrapped requests.exceptions.ConnectionError (as seen by
the customer).

This re-adds the retry on ConnectionError in addition to the internal
Cognite SDK classes, to be more resilient.