Lock when synchronization the state store (#275)
* Lock when synchronization the state store

We encountered:

> Unexpected error while synchronizing state store: dictionary changed
  size during iteration.

In this case, it is from an extractor with its own statestore that
builds upon the internal one, which may make this worse, but when
reviewing the code it seems we did not lock when inserting the state
to raw. This was probably the cause for the above error. In the same
way, we were serializing it in the JSON backed state store without
locking the underlying dict, which may in theory also cause this.

If we are concerned with long lock times around the insert, we could
make a lock, copy the state store, unlock and then insert/iterate over
that copy. It is a bit unclear if that is really needed.

* Fix init on start

---------

Co-authored-by: Mathias Lohne <mathias.lohne@cognite.com>