Fix a race condition resulting in broken state stores (#325)
Daemon threads do not block the program from exiting. This means that if
the state store was in the middle of syncing while the extractor shut
down, it would stop writing mid-way through leaving a broken json file.

Turning of daemon mode if a cancellation token was provided should fix
this issue. I also noticed we actually never set the cancellation token
for any state store when using the base class, so I fixed that too.