Improve file retries
Include external ID when logging failed files.

I also discovered an issue with how the retries was set up which caused
the uploads to not properly retry. The internal try/catch needed for the
thread pool ended up stopping retries. I split it out so that we have a
'wrap' function which applies the final layer of error handling.