Fix an issue with the retry decorator (#330)
By checking for cancellation before even calling the function, the retry
decorator was blocking final uploads in e.g. upload queues from even
starting.

This resulted in a very obscure bug where functions with retries would
seemingly work after the cancellation token was set, but in reality the
call stopped in the retry decorator and `None` was returned.