Add error reporting to integrations POC (#396)
Add an `error` method to the `Extractor` class that starts tracking of
an error. This can be used in one of two ways, either by manually
starting and stopping the error state:

``` python
e = extractor.error(...)

 # handle error

e.finish()
```

or by using it as a context

``` python
with extractor.error(...):
    # Handle error
```

You can create an instant error (with no duration) by using the
`instant()` method:

``` python
extractor.error(...).instant()
```

Tracking of start/end times, generating and keeping track of external
IDs, reporting in checkins, and so on is all handled automatically.