Clean up type hints in tests
Many tests were completely untyped, many had incorrect typing.

This doesn't fix everything, there's still a lot of type errors in the
tests, but it is an improvement. I'll maybe make a few more PRs cleaning
up a bit more in the tests, but some of it is also fine to leave as it
might get replaced altogether in the next major release.